Bug #57529: filter xattrs during sysvolsync
Since updating from Kernel 4.19 to Kernel 5.10 the behavior of xfs
seems to have changed with respect to the handling of xattrs.
As a symptom, rsync -AX seems to remove trusted.SGI_ACL_FILE when
synchronizing from the sysvol from a system with an ext4 partition.
Looks similar to https://github.com/RsyncProject/rsync/issues/301
and the workaround listed there almost works, but it caused problems
in combination with --delete-excluded, which then removed all "excluded"
xattrs (i.e. everything except security.NTACL) on the recipient side.
According to the man page, the use of that --delete-excluded option
modifies the behavior of --filter to only apply server side, but not
recipient side any longer. The documentation of the --xattrs option
mentions that under certain circumstances, the --filter rules need
to be explicitly extended, so now we also add the xattr exclude filter
rule explicitly for the recipient side.

univention/dev-issues/dev-incidents#67